Fixed GUI to offer more functionallity. This was a long needed change, glad that I finally organized myself to do it.
Started writing code for a hybrid prover/g/k-mod decomposition algorithm for producing minimal relations for the cone condition. I plan to prototype it quickly and get a hooked up version up and running (without any speed optimizations) by the end of the coming weekend, on a particular subalgebra k.