Simplified the multitasking by letting a single mutex control each control flow. Currently, there are two mutexes: one in ParallelComputing, which is global and can pause everything. The second is owned by CombinatorialChambersContainer.
The new idea of multitasking is: if you want to pause a thread, you simply lock a mutex and wait in an infinite cycle to to check the variable ParallelComputing::IsRunning. 

Note: the wait in an infinite cycle can be substituted with the use of conditional wait. Will do it in the future. For the time being I will keep it sweet and simple.

The mutexes have been implemented properly (hopefully) only on Linux. Gotta do it on Windows too.