Did a combination of math+computer work. Now the ElementOfUniversalEnveloping are allowed to have polynomial coefficients, *and* formal polynomial exponents.
The simplify function turned out to be actually simpler (due to me doing some math and making the formulas look& feel nice). The simplify function however will now fail to simplify if the exponent is not integral and the generator in question is on the right of the generator we need to commute with.

More or less, now I have implemented simplification on the condition that 1) the number of summands we get by commuting the generators with their formal powers is a constant (as a polynomial in the variables making up the formal exponents) and 2) the formal exponent around which we commute is on the right of the thing we want to commute it with.

The mathematical change of course carried with it a serious expansion of the Parser code (no conceptual changes though, of this I am very proud - my framework so far looks rock solid).