Started refactoring the root class to a templated class with the element type as a template argument.
This is the biggest code refactoring since merging the Rational and the LargeRational classes. 

Practically the entire project is affected. It will take a while to test everything and make sure works fine after the change. Not all projects should compile & link, however the main development project (the calculator.cbp, i.e. the new web server) does compile *and* link.

This commit is officially declared as "unstable".