Templated-in the coefficient ring of class ElementOfUniversalEnveloping.
Added function irreducibleRep to the calculator. As Littelmann paths give a monomial basis for the representation, as I am beginning to understand from P. Littelmann, "Cones, Crystals and patterns", this function will have to be simplified. At the moment the function spends 99% of its time verifying by direct computation what appears to be proven in section 10 of the cited article, so the whole thing will greatly simplify.