Added an experimental function for splitting tensor products of generalized Verma modules and finite dimensional modules.
I am not sure to what extent the operation I implemented actually does what the name of the function claims - I do not completely understand the mathematics behind it. 
In my defence, I am not sure that mathematics is at all available (in a reasonable form) from the existing literature. 
So, for what this mathematically experimental commit is good, here it goes.