Add openssl code to the build.
I need to peek into some of the internals of openssl
in order to work on the crypto library of the calculator.

The ultimate goal of the calculator's crypto library is to
have a self-contained implementation that is detached from openssl,
so I can get rid of the dependency.

I tried to peek non-invasively into openssl's internals but its
macros make it incredibly difficult to work with openssl
so I need to remove most of them.
Furthermore, in my opinion, anything but the most minimal use
of macros in code is a grave security risk and should be avoided at
all costs, which does not seem to be the philosophy of openssl.

Removing these macros is a lot of work, so I can't keep my work away from git, hence
I am creating the present branch.

I would also like to get rid of the openssl build system
which appears to use 3-4 layers of generated code
which, in my opinion, is a terrible design decision.
I personally find the build system boilerplate
to be more difficult to read and comprehend than the cryptographic
internals and algorithms, so it is a shame that
the technical difficulty of working with the code
is artificially created by unnecessary bloated boilerplate.