When the email does not match the username during login, use the username.
Previously we were failing the login instead, which is not ideal as
due to bugs in our code we have acquired an email that belongs to two accounts
and serves as the username of one of the accounts.
In particular, this means that the account in question would be locked out
of login.