Split the centralizer type computation into computation of individual simple components.
This should fix the bugs with the computation of the Dynkin index.
Also in type B_5, the centralizer of A_1^6 is incorrectly computed to be A_2^6,
this should be fixed as well.

The code has not been tested yet, this is a first version.
I will fix issues as I test this further in follow-ups.