Numerous bug fixes in the simple subalgebra component computation.
The current sl(2)-centralizer computation has an obvious bug in the dynkin type computation.

The smallest example where this shows is type C_3 where the centralizer of A_1^3
is reported to be of dimension 3 but its type is computed to A_2^8
(which is of dimension 2+2*3=8), which is clearly impossible.
I suspect I am not choosing the semisimple generating element correctly
and am not catching this correctly. Note the same produces
correct centralizers in F_4, which supports my hypothesis above.
I will double-check the math and follow up with a fix, hopefully soon.