Fixed the newly implemented AlgebraicNumber::evaluatesToComplex()
This was the root cause of the wrong sl(2)-centralizer computation. When selecting
positive roots, I was checking for the sign of the imaginary value of the
eigenvalue of the generic semisimple element used to split the root spaces of
the centarlizer. I was computing sqrt(negative) which was returning nan.
In turn, as I found out, nan>0 evaluates to true, so
all roots of the centralizer were being marked as positive, which
broke the simple type computation in an unusual way.