Vector change in basis fixed to work over an abitrary field.
The previous implementation required an invertible gram matrix
with respect to a symmetric scalar product; gram matrices
require hermitean scalar product to be invertible. This is
both difficult to compute with arbitrary algebraic numbers and
doesn't generalize well to finite characteristic.
All of this complexity is not needed, change of basis is just
linear algebra done properly. I've had the functions to do that
for years, but never fixed the implementation; better late than never.